requirements.txt
img2table
opencv-python-headless
wheel
pandas

#tesserocr
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
